+++ src/arch/loongarch32r/kernel/mysoc.c
@@ -11,7 +11,7 @@ void arch_time_init()
 	constant_timer_init();
 }
 
-void arch_mm_init(uint *pmem_base, uint *pmem_size)
+void arch_mm_init(ulong *pmem_base, ulong *pmem_size)
 {
 	*pmem_base = PMEM_BASE;
 	*pmem_size = PMEM_SIZE;
